+++ tests/wasm/simple.js
@@ -110,3 +110,39 @@ exports.test_string_roundtrip = () => {
   test('a longer string');
   test('a longer 💖 string');
 };
+
+exports.test_raw_pointers = function() {
+  const memory32 = new Uint32Array(wasm.__wasm.memory.buffer);
+  const memory8 = new Uint8Array(wasm.__wasm.memory.buffer);
+
+  const ptr1 = wasm.simple_return_raw_pointer_u32(4294967295);
+  assert.strictEqual(memory32[ptr1 / 4], 4294967295);
+  const ptr2 = wasm.simple_return_raw_pointer_u8(42);
+  assert.strictEqual(memory8[ptr2], 42);
+
+  wasm.simple_raw_pointers_work(ptr1, ptr2);
+  assert.strictEqual(memory32[ptr1 / 4], 42);
+  
+  const ptr3 = wasm.simple_return_raw_pointer_u32(4294967295);
+  wasm.simple_option_raw_pointers_work(ptr3, ptr2);
+  assert.strictEqual(memory32[ptr3 / 4], 42);
+
+  assert.strictEqual(wasm.simple_option_raw_pointers_work(0, ptr2), undefined);
+  assert.strictEqual(wasm.simple_option_raw_pointers_work(null, ptr2), undefined);
+  assert.strictEqual(wasm.simple_option_raw_pointers_work(undefined, ptr2), undefined);
+
+  assert.strictEqual(wasm.simple_option_raw_pointers_work(ptr1, 0), undefined);
+  assert.strictEqual(wasm.simple_option_raw_pointers_work(ptr1, null), undefined);
+  assert.strictEqual(wasm.simple_option_raw_pointers_work(ptr1, undefined), undefined);
+
+  assert.strictEqual(wasm.simple_return_option_null_pointer(), 0)
+};
+
+exports.test_non_null = function() {
+  assert.strictEqual(wasm.simple_option_nonnull_work(0), undefined);
+  assert.strictEqual(wasm.simple_option_nonnull_work(null), undefined);
+  assert.strictEqual(wasm.simple_option_nonnull_work(undefined), undefined);
+
+  assert.strictEqual(wasm.simple_option_nonnull_work(wasm.simple_return_non_null()), 42);
+  assert.strictEqual(wasm.simple_option_nonnull_work(wasm.simple_return_option_non_null(43)), 43);
+};